backend/requirements.txt
django
djangorestframework
django-cors-headers
djoser
pillow
exifread
celery
redis
gunicorn